feat(admin): move user management to /admin/users
The Admin sidebar item gains Overview and Users sub-items. /admin is
now the overview page and the user table lives at /admin/users under
the title Users. Each role form carries an id and its select and
button an accessible name, so a person's role can be changed by name.
